#81 refactor : 도전과제 삭제후 낙관적 업데이트 안되는 문제
